Move MilestoneReleaseStatus to constants module
