+++ rest-model/src/main/java/org/jboss/pnc/rest/restmodel/causeway/ReleaseStatus.java
@@ -17,7 +17,7 @@
  */
 package org.jboss.pnc.rest.restmodel.causeway;
 
-import org.jboss.pnc.model.MilestoneReleaseStatus;
+import org.jboss.pnc.enums.MilestoneReleaseStatus;
 
 /**
  * Author: Michal Szynkiewicz, michal.l.szynkiewicz@gmail.com
